refactor: 달력 모드에서만 사용하는 element의 파일 이름을 변경
